[Bug 3602264] Make clusterlib portable for 32-bit and 64-bit compilation.
git-svn-id: svn+ssh://svn.corp.yahoo.com/yahoo/yst/websearch/projects/clusterlib/trunk@253628 73f437aa-624d-0410-a3ad-9285e900ec26

+++ src/cli/cliparams.h
@@ -171,7 +171,7 @@ class CliParams {
      */
     void addToKeySet(const std::string &key)
     {
-        if (m_keySet.size() < m_keySetMaxSize) {
+        if (static_cast<int64_t>(m_keySet.size()) < m_keySetMaxSize) {
             m_keySet.insert(key);
         }
     }
